+++ R/zzz.R
@@ -28,12 +28,14 @@ NULL
   x = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
   x$task_col_roles$classif = unique(c(x$task_col_roles$classif, "always_included"))
   x$task_col_roles$regr = unique(c(x$task_col_roles$regr, "always_included"))
+  x$loaded_packages = c(x$loaded_packages, "mlr3fselect")
 
   # callbacks
   x = utils::getFromNamespace("mlr_callbacks", ns = "mlr3misc")
   x$add("mlr3fselect.backup", load_callback_backup)
   x$add("mlr3fselect.svm_rfe", load_callback_svm_rfe)
   x$add("mlr3fselect.one_se_rule", load_callback_one_se_rule)
+  x$add("mlr3fselect.internal_tuning", load_callback_internal_tuning)
 
   assign("lg", lgr::get_logger("bbotk"), envir = parent.env(environment()))
   if (Sys.getenv("IN_PKGDOWN") == "true") {