+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3fselect (development version)
+
 # mlr3fselect 1.1.1
 
 * compatibility: bbotk 1.1.1